Update patch set 4
Patch Set 4:

PRODUCT_DEVICE was overriden in PS2, I removed it in PS3, so now it's exactly the same as CM, just with an addition of newer fingerprint.

I think that clobbered build with this change will fix audio issue, as it's the same in CM, but I'm currently a bit out of time so this must wait for a few days longer.

Until then, feel free to test it.

Patch-set: 4
Label: Code-Review=0