Update patch set 3
Patch Set 3:

Yeah, the full_*.mk files are not used by Omni. Was discussing this on IRC the other day, I think initially the idea was to keep them around so that people could easily build AOSP but from what I understand that goal was dropped since AOSP will not build for most devices without a bit of modification anyway.

Still, it might make sense to split that off into its own commit, since it's not really related to this.

Patch-set: 3