Update patch set 2
Patch Set 2:

I am also not sure if it is audio hal or something else.
I have asked a friend of mine to test if the drain is there with stock audio hal.
BTW, if it is audio hal, I am highly suspicious it is the mics. They are the only things which may be left open without user knowing.
Can anyone do a test ? Reboot and record something via some app & test for drain issue.

Patch-set: 2