Update patch set 1
Patch Set 1:

> fixes the gr_text not defined, as previously it was passing a
 > variable as bold like gr_text_impl is supposed to. Now uses correct
 > gr_text format

This was added due to the updates to minui in AOSP not being in TWRP minui which are needed for healthd and broke the build. This was taken from marcos changes to the n80x device tree (https://gerrit.omnirom.org/10405/). There is actually a version of this in the smdk4412-common tree which needs to be taken care of, when this is resolved. In the meantime TWRP minui has been updated to provide a stub implementation of gr_clear() and everything compiles fine without these overrides. As of know lpm charging seems to be broken either way and I'm still in the process of figuring this out.

Patch-set: 1
Reviewer: Gerrit User 1001140 <1001140@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0