Update patch set 3
Patch Set 3:

> So the 3 PS apply fine and it compiles but my camera crashes. Are
 > these related camera patches supposed to fix that or is the
 > crashing another issue?

If it's compiling fine I suppose you also picked https://gerrit.omnirom.org/11530/.

For me the this is working mostly fine with the propriertary camera lib. Only front camera video recording is crashing from time to time. So basically the same way as it was in KK.

Patch-set: 3
Label: Code-Review=0