Update patch set 1
Patch Set 1:

In the last official builds libsensots was built, but overwritten - the vendor repo is not up to date.

The Accelerometer is working fine. If you're talking about the Linear Accelerometer: Yes, it showed the same misbehaviour due to the timestamps since it's a fused sensor and uses the same gyro integration to determine the rotational offset of the device to the earth surface.

Regarding the LUX meter: In my opinion it's working as good as one can expect from a sensor with such small surface. I compared it to two other devices (z3 compact and htc onex) and the absolute values are different but the relative values seem same to me. Maybe pauls formula can be slightly tweaked, but since I don't have any trustworthy reference readings I didn't change anything. Actually, if you look at the datasheet the values reported for white are supposed to be SI Lux values already.

Patch-set: 1