Update patch set 2
Patch Set 2:

As mentioned earlier, I have tested and all of the sensors were working fine. Probably you were not printing the values properly. Try to check the actual sysfs values to see if they are reflecting properly. set_delay will fail if the sensor is not enabled. I have tested the delays of upto 10seconds and they work fine.

It's not a common and a good practice to use strtoimax because intmax_t depends on the compiler and the target platform. It only makes sense to use it when you are dealing with numbers which are very large and you are not sure about the size. strtoll on the other hand returns long long int, which is exactly what we need.

Patch-set: 2
Reviewer: Gerrit User 1000280 <1000280@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0