Update patch set 2
Patch Set 2:

> As mentioned earlier, I have tested and all of the sensors were
 > working fine. Probably you were not printing the values properly.
 > Try to check the actual sysfs values to see if they are reflecting
 > properly. set_delay will fail if the sensor is not enabled. I have
 > tested the delays of upto 10seconds and they work fine.
 
Haha, no, I know how to print the values properly - don't worry ;)
The problem was actually the compiler cache: I didn't clean my dir after applying the original version of this patch. Therefore it seems the sensor-specific files were not rebuilt and in consequence the implicit conversion to int64_t was not inserted, which lead to the errors described.

Patch-set: 2
Label: Code-Review=0