Update patch set 3
Patch Set 3:

Regarding strtoimax vs stroll:
You were the one who wanted this to be as cleanly platform-independet as possible. So using stroll() will again assume platform-specific properties. Even though long long is to be guaranteed to be at least 64bit, it could also be bigger. Using strtoimax is the same scenario: We can be sure it's at least as big as int64_t but could well be bigger. So this basically just a question of style - I chose strtoimax since it avoids the mentioning of LL.
Same for the cast - in both cases the result on our plattform will be long long and therefore the cast will be optimized away anyway, it just serves as an indicator as to what we expect. On another plattform with long long being e.g. 128bit and in consequence intmax being bigger too, this would lead to a possible overflow, but since we know our values will alway fit int64_t it is guaranteed to be okay.

As I said before: By that logic we could just kept using long long since - as you just said - it is exactly what we want.

Patch-set: 3